Add yq to tools image, use image in make, add update_container_documentation to check.sh

+++ bin/check.sh
@@ -12,6 +12,15 @@ if [ ! -z "$(git status --porcelain $requirements_files)" ];then
   fail=1
 fi
 
+documentation_file="documentation/Docker-container-profiles.md"
+bin/update_container_documentation.sh
+if [ ! -z "$(git diff --word-diff=porcelain -G'^\| ' $documentation_file)" ];then
+  echo -e "\e[31mThe docker container documentation is not up to date, please run 'make update_container_documentation' and commit the documentation to Git!"
+  git diff --word-diff=porcelain -G'^\| ' $documentation_file
+  fail=1
+fi
+
+
 # verify DEBUG can only be enabled when authentication is set (should exit 1 and print error message)
 command="docker run -ti --rm -e DEBUG=True -e AUTH_ALL_URLS -e ALLOW_LIST ghcr.io/internetstandards/webserver"
 output=$($command)
